+++ RobotProject/src/pl/poznan/put/ioiorobot/widgets/MapWidget.java
@@ -93,8 +93,8 @@ public class MapWidget extends View {
 	}
 
 	public void addPosition(Position pos) {
-		double yParam = -pos.y() + 200;
-		double xParam = -pos.x();
+		float yParam = -pos.y() + 200;
+		float xParam = -pos.x();
 
 		position.set(xParam, yParam, pos.angle());
